tests(Performance): update base number first time since 2023
Execution if the tests were skipped for a long time as the condition
included to run on server master and PHP 8.2 which did not happen for a
long time. Keeping the value at the lowest master version so we notice
when to bump. Also a lot of tests were added since, so updating the base
number as the old one has no meaning anymore.

Signed-off-by: Arthur Schiwon <blizzz@arthur-schiwon.de>